Tests/LibWeb/Text/expected/hit_testing/content-editable-toggle.txt

caret before toggle: null
caret with contenteditable: [object HTMLDivElement] @0
caret after removal: null
caret with designMode: [object HTMLDivElement] @0
caret after designMode off: null
